Adjust folder and rename interface #18
