+++ Config/DefaultModularFeatures_ExtraActions.ini
@@ -1,0 +1,2 @@
+[CoreRedirects]
++ClassRedirects=(OldName="/Script/ModularFeatures_ExtraActions.AbilityInputBinding", NewName="/Script/ModularFeatures_ExtraActions.MFEA_AbilityInputBinding")
