+++ ModularFeatures_ExtraActions.uplugin
@@ -1,7 +1,7 @@
 {
 	"FileVersion": 3,
 	"Version": 1,
-	"VersionName": "1.2.0",
+	"VersionName": "1.2.1",
 	"FriendlyName": "Modular Features: Extra Actions",
 	"Description": "Integrates GAS and Enhanced Input to the existing Game Features and Modular plugins.",
 	"Category": "Game Features",
