+++ Source/Private/Actions/GameFeatureAction_AddAbilities.cpp
@@ -2,7 +2,7 @@
 // Year: 2022
 // Repo: https://github.com/lucoiso/UEModularFeatures_ExtraActions
 
-#include "GameFeatureAction_AddAbilities.h"
+#include "Actions/GameFeatureAction_AddAbilities.h"
 #include "Components/GameFrameworkComponentManager.h"
 #include "InputAction.h"
 #include "ModularFeatures_InternalFuncs.h"
@@ -122,7 +122,7 @@ void UGameFeatureAction_AddAbilities::AddActorAbilities(AActor* TargetActor, con
 			if (!Ability.InputAction.IsNull())
 			{
 				// Get the target Ability Input Binding interface
-				const IAbilityInputBinding* const SetupInputInterface = ModularFeaturesHelper::GetAbilityInputBindingInterface(TargetActor, InputBindingOwnerOverride);
+				const IMFEA_AbilityInputBinding* const SetupInputInterface = ModularFeaturesHelper::GetAbilityInputBindingInterface(TargetActor, InputBindingOwnerOverride);
 
 				// If we can bind the input to the target interface, we must add the input reference to the ability data
 				if (UInputAction* const AbilityInput = Ability.InputAction.LoadSynchronous(); 
@@ -181,7 +181,7 @@ void UGameFeatureAction_AddAbilities::RemoveActorAbilities(AActor* TargetActor)
 		}
 
 		// Get the interface owner and try to remove the input bindings
-		if (const IAbilityInputBinding* const SetupInputInterface = ModularFeaturesHelper::GetAbilityInputBindingInterface(TargetActor, InputBindingOwnerOverride))
+		if (const IMFEA_AbilityInputBinding* const SetupInputInterface = ModularFeaturesHelper::GetAbilityInputBindingInterface(TargetActor, InputBindingOwnerOverride))
 		{
 			ModularFeaturesHelper::RemoveAbilityInputInInterfaceOwner(SetupInputInterface->_getUObject(), ActiveAbilities.InputReference);
 		}
