+++ Source/Public/Actions/GameFeatureAction_AddEffects.h
@@ -7,7 +7,7 @@
 #include "CoreMinimal.h"
 #include "GameplayTagContainer.h"
 #include "GameplayEffectTypes.h"
-#include "GameFeatureAction_WorldActionBase.h"
+#include "Actions/GameFeatureAction_WorldActionBase.h"
 #include "GameFeatureAction_AddEffects.generated.h"
 
 class UGameplayEffect;
